Refactor: 파일 구조 변경
- 각 Entity마다 독립적인 디렉터리를 갖도록 변경
