+++ wekiki/src/main/java/com/kakao/techcampus/wekiki/comment/Comment.java
@@ -1,9 +1,7 @@
-package com.kakao.techcampus.wekiki.page.history;
+package com.kakao.techcampus.wekiki.comment;
 
-import com.kakao.techcampus.wekiki.group.member.GroupMember;
 import com.kakao.techcampus.wekiki.member.Member;
-import com.kakao.techcampus.wekiki.page.Page;
-import com.kakao.techcampus.wekiki.page.post.Post;
+import com.kakao.techcampus.wekiki.post.Post;
 import jakarta.persistence.*;
 import lombok.AccessLevel;
 import lombok.Builder;
@@ -15,26 +13,23 @@ import java.time.LocalDateTime;
 @Getter
 @Entity
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
-@Table(name = "history_tb")
-public class History {
+@Table(name = "comment_tb")
+public class Comment {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
     @ManyToOne
-    private GroupMember groupMember;
-    @ManyToOne
-    private Page page;
+    private Member member;
     @ManyToOne
     private Post post;
     private String content;
     private LocalDateTime created_at;
 
     @Builder
-    public History(Long id, GroupMember groupMember, Page page, Post post, String content, LocalDateTime created_at) {
+    public Comment(Long id, Member member, Post post, String content, LocalDateTime created_at) {
         this.id = id;
-        this.groupMember = groupMember;
-        this.page = page;
+        this.member = member;
         this.post = post;
         this.content = content;
         this.created_at = created_at;
