+++ wekiki/src/main/java/com/kakao/techcampus/wekiki/history/History.java
@@ -1,7 +1,8 @@
-package com.kakao.techcampus.wekiki.page.post.comment;
+package com.kakao.techcampus.wekiki.history;
 
-import com.kakao.techcampus.wekiki.member.Member;
-import com.kakao.techcampus.wekiki.page.post.Post;
+import com.kakao.techcampus.wekiki.group.member.GroupMember;
+import com.kakao.techcampus.wekiki.page.Page;
+import com.kakao.techcampus.wekiki.post.Post;
 import jakarta.persistence.*;
 import lombok.AccessLevel;
 import lombok.Builder;
@@ -13,23 +14,26 @@ import java.time.LocalDateTime;
 @Getter
 @Entity
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
-@Table(name = "comment_tb")
-public class Comment {
+@Table(name = "history_tb")
+public class History {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
     @ManyToOne
-    private Member member;
+    private GroupMember groupMember;
+    @ManyToOne
+    private Page page;
     @ManyToOne
     private Post post;
     private String content;
     private LocalDateTime created_at;
 
     @Builder
-    public Comment(Long id, Member member, Post post, String content, LocalDateTime created_at) {
+    public History(Long id, GroupMember groupMember, Page page, Post post, String content, LocalDateTime created_at) {
         this.id = id;
-        this.member = member;
+        this.groupMember = groupMember;
+        this.page = page;
         this.post = post;
         this.content = content;
         this.created_at = created_at;
