+++ wekiki/src/main/java/com/kakao/techcampus/wekiki/report/Report.java
@@ -1,4 +1,4 @@
-package com.kakao.techcampus.wekiki.page.report;
+package com.kakao.techcampus.wekiki.report;
 
 import com.kakao.techcampus.wekiki.group.member.GroupMember;
 import com.kakao.techcampus.wekiki.page.Page;
